requirements.txt
Flask
pandas
numpy
scikit-learn
ultralytics
Pillow
